Add is_solr_query filter
Right now it's impossible to turn off Solr for specific types of searches. 

Another considered solution was to allow for $query->get( 'solr_integrate' ); to be set to false and have that overwrite the result of $query->is_search() but the filter seems like it would give more flexibility and be clearer in it's goal.